==> 03-eda/3-evaluate-model.qmd <==
---
title: "Neural-net"
author: "Dusty Turner"
format: html
engines:
  R: true
  python: true
cache: true
warning: false
message: false
---

## Testing

```{r}
library(tidyverse)
# library(keras)
# library(tensorflow)
# library(reticulate)
# library(data.table)
library(gganimate)
# library(arrow)

source(here::here("03-eda","0-source-functions.R"))

test_joined_with_preds <- read_rds("04-models/test_joined_with_preds_2024-01-07_16-03-36.rds")

# week_1 <- read_parquet(here::here("02-clean-data", "week_1.parquet"))

test_joined_with_preds 

```



```{r}

test_joined_with_preds |> 
  count(binary_predictions)

test_joined_with_preds |> 
  yardstick::accuracy(truth = tackle, estimate = binary_predictions)
test_joined_with_preds |> 
  yardstick::recall(truth = tackle, estimate = binary_predictions)
test_joined_with_preds |> 
  yardstick::precision(truth = tackle, estimate = binary_predictions)

library(yardstick)
# Define the multiple classification metrics
classification_metrics <- metric_set(accuracy, precision, recall)


fake <- factor(0, levels = c(0,1))

test_joined_with_preds %>%
  mutate(fake_prediction = fake) |> 
  # filter(frames_from_tackle >= -40) %>%
  # group_by(frames_from_tackle, play_type) %>%
  classification_metrics(truth = tackle, estimate = fake_prediction)

# Calculate the metrics for your data
accuracy_plot <-
test_joined_with_preds %>%
  filter(frames_from_tackle >= -40) %>%
  group_by(frames_from_tackle, play_type) %>%
  classification_metrics(truth = tackle, estimate = binary_predictions) |>
  mutate(.metric = str_to_title(.metric)) |> 
  mutate(play_type = str_to_sentence(play_type)) |>
  mutate(play_type = fct_relevel(play_type, "Run")) |> 
  ggplot(aes(x = frames_from_tackle, y = .estimate, alpha = frames_from_tackle)) +
  geom_point() +
  geom_line() +
  scale_y_continuous(labels = scales::percent_format(accuracy = 0.1, scale = 100)) +
  scale_x_continuous(labels = c(4,3,2,1,0), breaks = c(-40,-30,-20,-10,0)) +
  facet_grid(.metric~play_type, scales = "free_y") +
  # facet_grid(play_type~.metric, nrow = 3, scales = "free_y") +
  labs(y = "", x = "Seconds From Tackle") +
  theme_minimal(base_size = 13) +
  theme(legend.position = "none", panel.grid.minor = element_blank(), 
        panel.background = element_rect(color = "grey75"), 
        plot.background = element_rect(fill = "white", color = NA)) 

ggsave("05-presentation-files/accuracy-plot-1000-400.svg", plot = accuracy_plot, width = 9, height = 6)

```

```{r}

library(pROC)

# Creating the ROC object
roc_result<- roc(response = test_joined_with_preds$tackle, predictor = test_joined_with_preds$final_predictions)

# Calculating the AUC
auc_value <- auc(roc_result)

# Plotting the ROC curve
ggplot(data = data.frame(fpr = 1 - roc_result$specificities, tpr = roc_result$sensitivities), aes(x = fpr, y = tpr)) +
  geom_line() +
  geom_abline(linetype = "dashed") +
  labs(x = "False Positive Rate", y = "True Positive Rate", title = "ROC Curve") +
  annotate("text", x = 0.6, y = 0.4, label = paste("AUC =", round(auc_value, 2)))


```


## Animations

```{r}
library(gganimate)

test_joined_with_preds |> 
  filter(play_type == "scramble") %>% 
  count(game_idplay_id)

# week_1 %>% filter(game_idplay_id == "20220908001102") |> filter(tackle == 1) |> select(display_name, jersey_number)
# week_1 %>% filter(game_idplay_id == "20220908001030") |> filter(tackle == 1) |> select(display_name, jersey_number)
# week_1 %>% filter(game_idplay_id == "20220908001358") |> filter(tackle == 1) |> select(display_name, jersey_number)
week_1 %>% filter(game_idplay_id == "20220911001587") |> filter(tackle == 1) |> select(display_name, jersey_number)

# Assuming week_1 is your data frame
animated_plot <-
  # week_1 %>% filter(game_idplay_id == "20220908001102") |> 
  # week_1 %>% filter(game_idplay_id == "20220908001030") |> 
  # week_1 %>% filter(game_idplay_id == "20220908001358") |> 
  week_1 %>% filter(game_idplay_id == "20220911001587") |> 
  left_join(test_joined_with_preds %>% select(game_id, play_id, display_name, frame_id, final_predictions), by = c("game_id", "play_id", "display_name", "frame_id")) %>% 
  filter(!is.na(final_predictions)) %>% 
  group_by(game_id, play_id) %>% 
  # filter(cur_group_id() == 4) |> 
  ungroup()  |> 
  select(frame_id, final_predictions, jersey_number) |> 
  ggplot(aes(x = frame_id, y = final_predictions)) + 
  geom_line() +
  geom_point(show.legend = FALSE, size = 3) + 
  facet_wrap(~jersey_number) +
  transition_reveal(frame_id)

anm1 <- animate(animated_plot, width = 800, height = 600, nframes = 100)


animated_plot2 <-
week_1 |> 
  # filter(game_idplay_id == "20220908001102") |>
  # filter(game_idplay_id == "20220908001030") |>
  # filter(game_idplay_id == "20220908001358") |>
  filter(game_idplay_id == "20220911001587") |>
  left_join(test_joined_with_preds) |> 
  relocate(final_predictions) |> 
  filter(frame_id <= max(frame_id)-4) |> 
  mutate(final_predictions = ifelse(club == defensive_team & is.na(final_predictions), 0, final_predictions)) |> 
  mutate(color = final_predictions) |> 
  mutate(jersey_number = ifelse(defensive_team == club, jersey_number, "")) |> 
  select(display_name, distance_to_ball, x, y, color, absolute_yardline_number, ball_carrier, play_id, time, play_description, is_football, club, jersey_number, final_predictions, defensive_team, frame_id) %>%
  # select(frame_id, display_name, final_predictions) |> print(n = Inf)
{
  ggplot(data = ., aes(x = x, y = y, color = color)) +
      # ggplot(aes(x = x, y = y, color = color)) +
  geom_vline(aes(xintercept = absolute_yardline_number), color = "blue") +
  geom_point(aes(shape = is_football), size = 3, show.legend = FALSE) +
  geom_text(aes(label = jersey_number), color = "black", nudge_y = -1) +
  scale_color_gradient(low = "grey", high = "black", na.value = "dodgerblue") +
  transition_time(time) + ease_aes("linear") +
  labs(y = "", x = "Yards To Endzone", title = str_wrap(.$play_description[1], width = 80)) +
  theme_field
}

anm2 <- animate(animated_plot2, width = 800, height = 600, nframes = 100)

animated_plot3 <-
week_1 |> 
  # filter(game_idplay_id == "20220908001102") |>
  filter(game_idplay_id == "20220908001030") |>
  # select(tackle, jersey_number) |> print(n = Inf)
  left_join(test_joined_with_preds) |> 
  relocate(final_predictions) |> 
  filter(frame_id <= max(frame_id)-4) |> 
  mutate(final_predictions = ifelse(club == defensive_team & is.na(final_predictions), 0, final_predictions)) |> 
  group_by(frame_id) |> 
  mutate(max_predict = final_predictions == max(final_predictions, na.rm = TRUE)) |> 
  ungroup() |> 
  mutate(color = max_predict) |> 
  mutate(jersey_number = ifelse(defensive_team == club, jersey_number, "")) |> 
  select(max_predict, display_name, distance_to_ball, x, y, color, absolute_yardline_number, ball_carrier, play_id, time, play_description, is_football, club, jersey_number, final_predictions, defensive_team, frame_id) %>% 


{
  ggplot(data = ., aes(x = x, y = y, color = color)) +
      # ggplot(aes(x = x, y = y, color = color)) +
  geom_vline(aes(xintercept = absolute_yardline_number), color = "blue") +
  geom_point(aes(shape = is_football), size = 3, show.legend = FALSE) +
  geom_text(aes(label = jersey_number), color = "black", nudge_y = -1) +
  # scale_color_gradient(low = "grey", high = "black", na.value = "dodgerblue") +
  transition_time(time) + ease_aes("linear") +
  labs(y = "", x = "Yards To Endzone", title = str_wrap(.$play_description[1], width = 80)) +
  theme_field
}

anm3 <- animate(animated_plot3, width = 800, height = 600, nframes = 100)

# nn_briar <- 
  test_joined_with_preds |> 
  select(game_id, play_id, display_name,tackle, final_predictions) |> 
  filter(!is.na(final_predictions)) |> 
  mutate(tackle  = as.integer(as.character(tackle))) |> 
  group_by(game_id, play_id, display_name) |> 
  reframe(expected_prob_of_tackle = mean(final_predictions), tackle = mean(tackle)) |> 
  mutate(tackles_over_expected_play = ifelse(tackle == 1, 1-expected_prob_of_tackle, -expected_prob_of_tackle)) |> 
  group_by(display_name) |> 
  reframe(tackles_over_expected = sum(tackles_over_expected_play)) |> 
  arrange(tackles_over_expected)


```
